Extract mapped staging allocation helpers for renderer uploads
Move the existing cache without changing allocation, ownership or reuse behavior. Ref #3407.
